Indri bug fix. Probably last major core Indri update for this project.

+++ indri/branches/metzler/include/indri/CachedFrequencyBeliefNode.hpp
@@ -55,8 +55,8 @@ public:
   const greedy_vector<bool>& hasMatch( int documentID, const greedy_vector<Extent>& extents ); 
   const std::string& getName() const;
   
-  int matches( int begin, int end );
-  int contextSize( int begin, int end );
+  double matches( int begin, int end );
+  double contextSize( int begin, int end );
   
   void advance();
 };
